+++ extensions/google-translate/src/QuickTranslate/LanguageDropdown.tsx
@@ -3,10 +3,13 @@ import { useSourceLanguage, useTargetLanguages } from "../hooks";
 import { LanguageCode, languages, supportedLanguagesByCode } from "../languages";
 import { TargetLanguageList } from "./TargetLanguageList";
 
-export function LanguageDropdown() {
+export function LanguageDropdown(props: { translatedSourceLanguage?: LanguageCode }) {
   const navigation = useNavigation();
   const [sourceLanguage, setSourceLanguage] = useSourceLanguage();
   const [targetLanguages] = useTargetLanguages();
+  const autoDetectedLanguageName = props.translatedSourceLanguage
+    ? supportedLanguagesByCode[props.translatedSourceLanguage].name
+    : null;
   return (
     <List.Dropdown
       value={sourceLanguage}
@@ -25,9 +28,21 @@ export function LanguageDropdown() {
         title={`Translate to  ->  ${targetLanguages.map((l) => supportedLanguagesByCode[l].name).join(" ")}`}
         value="manageTargetLanguages"
       />
-      {languages.map((lang) => (
-        <List.Dropdown.Item key={lang.code} title={lang.name} value={lang.code} />
-      ))}
+      {languages.map((lang) =>
+        lang.code === "auto" ? (
+          <List.Dropdown.Item
+            key={lang.code}
+            title={
+              autoDetectedLanguageName && sourceLanguage === "auto"
+                ? `${lang.name} (${autoDetectedLanguageName})`
+                : lang.name
+            }
+            value={lang.code}
+          />
+        ) : (
+          <List.Dropdown.Item key={lang.code} title={lang.name} value={lang.code} />
+        ),
+      )}
     </List.Dropdown>
   );
 }